refactor: inferrer move
