+++ modules/common/src/main/java/org/vstu/meaningtree/languages/TranslatorContext.java
@@ -14,7 +14,7 @@ import org.vstu.meaningtree.nodes.modules.Import;
 import org.vstu.meaningtree.nodes.statements.CompoundStatement;
 import org.vstu.meaningtree.utils.scopes.ScopeTable;
 import org.vstu.meaningtree.utils.scopes.ScopeTableElement;
-import org.vstu.meaningtree.utils.typeinference.SimpleTypeInferrer;
+import org.vstu.meaningtree.utils.scopes.SimpleTypeInferrer;
 
 import java.lang.reflect.Method;
 import java.util.*;
