+++ modules/common/src/main/java/org/vstu/meaningtree/utils/scopes/SimpleTypeInferrer.java
@@ -1,4 +1,4 @@
-package org.vstu.meaningtree.utils.typeinference;
+package org.vstu.meaningtree.utils.scopes;
 
 import org.jetbrains.annotations.NotNull;
 import org.vstu.meaningtree.nodes.*;
@@ -29,7 +29,6 @@ import org.vstu.meaningtree.nodes.statements.conditions.SwitchStatement;
 import org.vstu.meaningtree.nodes.types.UnknownType;
 import org.vstu.meaningtree.nodes.types.builtin.*;
 import org.vstu.meaningtree.nodes.types.containers.*;
-import org.vstu.meaningtree.utils.scopes.ScopeTable;
 
 import java.util.ArrayList;
 import java.util.List;
